feat: 好友房相关改动

+++ room/src/main/java/com/jjg/game/room/controller/AbstractRoomController.java
@@ -41,7 +41,7 @@ import java.util.stream.Collectors;
  * @date 2025/6/25 12:33
  */
 public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room> implements TimerListener<IProcessorHandler>,
-        IRoomLifeCycle {
+    IRoomLifeCycle {
     protected final Logger log = LoggerFactory.getLogger(getClass());
     //房间对象
     protected R room;
@@ -62,7 +62,7 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
     // 房间配置
     protected RC roomCfg;
     // 机器人上次创建的时间
-    private long robotLastCreatedTime;
+    protected long robotLastCreatedTime;
     // 房间状态
     private volatile ERoomState roomState;
 
@@ -78,10 +78,10 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
      */
     public <RD extends GameDataVo<RC>> AbstractGameController<RC, RD> createGameController(RC roomCfg) {
         Set<Class<? extends AbstractGameController<? extends RoomCfg, ? extends GameDataVo<? extends RoomCfg>>>> gameControllerClazz =
-                roomManager.getGameControllerClazz();
+            roomManager.getGameControllerClazz();
         try {
             for (Class<? extends AbstractGameController<? extends RoomCfg, ? extends GameDataVo<? extends RoomCfg>>> controllerClazz :
-                    gameControllerClazz) {
+                gameControllerClazz) {
                 GameController gameAnnotateController = controllerClazz.getAnnotation(GameController.class);
                 EGameType games = gameAnnotateController.gameType();
                 RoomType roomType = RoomType.getRoomType(roomCfg.getId());
@@ -89,7 +89,7 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
                 // 游戏类型和房间类型必须一致
                 if (games.getGameTypeId() == roomCfg.getGameID() && isEqualsRoomType) {
                     Constructor<AbstractGameController<RC, RD>> constructor =
-                            (Constructor<AbstractGameController<RC, RD>>) controllerClazz.getDeclaredConstructor(AbstractRoomController.class);
+                        (Constructor<AbstractGameController<RC, RD>>) controllerClazz.getDeclaredConstructor(AbstractRoomController.class);
                     // 将调用当前方法的RoomController写入GameController中
                     AbstractGameController<RC, RD> gameController = constructor.newInstance(this);
                     RD roomDataVoCopied = gameController.createRoomDataVo(roomCfg);
@@ -124,10 +124,10 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
                 if (!doResult.success()) {
                     if (!playerController.isRobotPlayer()) {
                         log.warn("加入房间失败 gameType = {},roomId = {},playerId = {} 加入结果：{}",
-                                playerController.getPlayer().getGameType(),
-                                this.room.getId(),
-                                playerController.playerId(),
-                                doResult.code);
+                            playerController.getPlayer().getGameType(),
+                            this.room.getId(),
+                            playerController.playerId(),
+                            doResult.code);
                     }
                     result.code = doResult.code;
                     return result;
@@ -166,14 +166,18 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
      * 更新房间等待列表
      */
     protected void updateWaitRoomList() {
+        // 好友房不进等待房间列表
+        if (room instanceof FriendRoom) {
+            return;
+        }
         // 如果房间已满，则需要将等待房间列表从redis中删除
         if (!room.canEnter()) {
             roomManager.getMatchDataDao().removeWaitJoinRoomId(
-                    room.getGameType(), room.getRoomCfgId(), room.getId());
+                room.getGameType(), room.getRoomCfgId(), room.getId());
         } else {
             // 如果房间未满，则将直接写入等待列表
             roomManager.getMatchDataDao().addWaitJoinRoomId(
-                    room.getGameType(), room.getRoomCfgId(), room.getId(), room.getCreateTime());
+                room.getGameType(), room.getRoomCfgId(), room.getId(), room.getCreateTime());
         }
     }
 
@@ -185,59 +189,59 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
     protected CommonResult<? extends Room> checkRoomCanJoin(PlayerController playerController) {
         R thatRoom = this.room;
         return roomDao.doSave(playerController.getPlayer().getGameType(),
-                this.room.getId()
-                , new DataSaveCallback<>() {
-                    @Override
-                    public void updateData(Room dataEntity) {
-                    }
+            this.room.getId()
+            , new DataSaveCallback<>() {
+                @Override
+                public void updateData(Room dataEntity) {
+                }
 
-                    @Override
-                    public Boolean updateDataWithRes(Room room) {
-                        try {
-                            if (room == null) {
-                                log.error("检查房间：{} 是否可以加入时, 房间为空，player: {}",
-                                        thatRoom.logStr(), playerController.playerId());
-                                return false;
+                @Override
+                public Boolean updateDataWithRes(Room room) {
+                    try {
+                        if (room == null) {
+                            log.error("检查房间：{} 是否可以加入时, 房间为空，player: {}",
+                                thatRoom.logStr(), playerController.playerId());
+                            return false;
+                        }
+                        //先检查该玩家是否已经在该房间中
+                        boolean exist = room.hasPlayer(playerController.playerId());
+                        if (exist) {
+                            log.debug("玩家已在房间中 gameType = {},roomId = {},playerId = {}",
+                                playerController.getPlayer().getGameType(), thatRoom.getId(),
+                                playerController.playerId());
+                            return true;
+                        }
+
+                        if (!room.canEnter()) {
+                            if (!playerController.isRobotPlayer()) {
+                                log.debug("该房间无法进入 gameType = {},roomId = {},playerId = {} roomPlayerSize: {} " +
+                                        "roomLimit: {}",
+                                    playerController.getPlayer().getGameType(), thatRoom.getId(),
+                                    playerController.playerId(),
+                                    AbstractRoomController.this.room.getRoomPlayers().size(),
+                                    AbstractRoomController.this.room.getMaxLimit());
                             }
-                            //先检查该玩家是否已经在该房间中
-                            boolean exist = room.hasPlayer(playerController.playerId());
-                            if (exist) {
-                                log.debug("玩家已在房间中 gameType = {},roomId = {},playerId = {}",
-                                        playerController.getPlayer().getGameType(), thatRoom.getId(),
-                                        playerController.playerId());
+                            return false;
+                        }
+
+                        //如果之前不在房间，则按座位排座
+                        for (int i = 0; i < room.getMaxLimit(); i++) {
+                            boolean flag = room.setHasPlayer(i);
+                            if (!flag) {
+                                RoomPlayer tmpRoomPlayer = roomDao.createRoomPlayer(playerController);
+                                tmpRoomPlayer.setSit(i);
+                                tmpRoomPlayer.setOnline(true);
+                                tmpRoomPlayer.setRobot(playerController.getPlayer() instanceof RobotPlayer);
+                                room.addPlayer(tmpRoomPlayer);
                                 return true;
                             }
-
-                            if (!room.canEnter()) {
-                                if (!playerController.isRobotPlayer()) {
-                                    log.debug("该房间无法进入 gameType = {},roomId = {},playerId = {} roomPlayerSize: {} " +
-                                                    "roomLimit: {}",
-                                            playerController.getPlayer().getGameType(), thatRoom.getId(),
-                                            playerController.playerId(),
-                                            AbstractRoomController.this.room.getRoomPlayers().size(),
-                                            AbstractRoomController.this.room.getMaxLimit());
-                                }
-                                return false;
-                            }
-
-                            //如果之前不在房间，则按座位排座
-                            for (int i = 0; i < room.getMaxLimit(); i++) {
-                                boolean flag = room.setHasPlayer(i);
-                                if (!flag) {
-                                    RoomPlayer tmpRoomPlayer = roomDao.createRoomPlayer(playerController);
-                                    tmpRoomPlayer.setSit(i);
-                                    tmpRoomPlayer.setOnline(true);
-                                    tmpRoomPlayer.setRobot(playerController.getPlayer() instanceof RobotPlayer);
-                                    room.addPlayer(tmpRoomPlayer);
-                                    return true;
-                                }
-                            }
-                        } catch (Exception e) {
-                            log.error("检查房间是否可以加入时发生异常", e);
                         }
-                        return false;
+                    } catch (Exception e) {
+                        log.error("检查房间是否可以加入时发生异常", e);
                     }
-                });
+                    return false;
+                }
+            });
     }
 
     /**
@@ -260,7 +264,7 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
                 e.getParameter().action();
             } catch (Exception ex) {
                 log.error("执行房间：{} 游戏类型：{} 定时器异常, msg: {} ",
-                        room.getId(), room.getGameType(), ex.getMessage(), ex);
+                    room.getId(), room.getGameType(), ex.getMessage(), ex);
             }
         }
     }
@@ -285,8 +289,8 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
     }
 
     @Override
-    public boolean continueGame() {
-        if (gameController.continueGame()) {
+    public boolean tryContinueGame() {
+        if (gameController.tryContinueGame()) {
             roomState = ERoomState.GAMING;
             return true;
         }
@@ -361,7 +365,7 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
         // 房间Timer执行tick时间 现在默认 200ms
         // 添加房间tick
         timerCenter.add(new RoomTimerEvent<>(
-                this, room, this::roomTick, RoomConstant.ROOM_TICK_TIME, RoomEventType.ROOM_TICK));
+            this, room, this::roomTick, RoomConstant.ROOM_TICK_TIME, RoomEventType.ROOM_TICK));
         roomState = ERoomState.READY;
     }
 
@@ -412,7 +416,7 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
         int exited = roomManager.exitRoom(playerController);
         if (exited != Code.SUCCESS) {
             log.info("换房间时退出当前房间失败 playerId:{} oldRoomId:{} gameType:{} roomConfigId:{}",
-                    oldRoomId, roomOtherId, gameType, roomConfigId);
+                oldRoomId, roomOtherId, gameType, roomConfigId);
             return false;
         }
         //加入房间
@@ -450,7 +454,7 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
         }
         // 创建一个机器人
         PlayerController robotPlayerController =
-                robotService.getOrCreateRobotPlayerController(roomCfgId, room.getId());
+            robotService.getOrCreateRobotPlayerController(roomCfgId, room.getId());
         if (System.currentTimeMillis() - robotCreateStartTime >= 200) {
             log.debug("机器人创建超时，花费时间：{}", System.currentTimeMillis() - robotCreateStartTime);
         }
@@ -489,17 +493,17 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
             playerControllers.remove(playerController.getPlayer().getId());
             //从room中移除
             R room = roomDao.removePlayer(
+                playerController.getPlayer().getGameType(),
+                playerController.roomId(),
+                playerController.playerId());
+            if (room != null) {
+                log.debug("强制离开房间成功, gameType = {},roomId = {},playerId = {}",
                     playerController.getPlayer().getGameType(),
                     playerController.roomId(),
                     playerController.playerId());
-            if (room != null) {
-                log.debug("强制离开房间成功, gameType = {},roomId = {},playerId = {}",
-                        playerController.getPlayer().getGameType(),
-                        playerController.roomId(),
-                        playerController.playerId());
             } else {
                 log.debug("将玩家从房间中移除失败 gameType = {},roomId = {},playerId = {}",
-                        this.room.getGameType(), this.room.getId(), playerController.playerId());
+                    this.room.getGameType(), this.room.getId(), playerController.playerId());
                 result.code = Code.FAIL;
                 return result;
             }
@@ -528,14 +532,14 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
             this.playerControllers.values().removeAll(playerControllers);
             //从room中移除所有玩家
             R room = roomDao.removePlayers(
-                    this.room.getGameType(),
-                    this.room.getId(),
-                    playerControllers.stream().map(PlayerController::playerId).toList());
+                this.room.getGameType(),
+                this.room.getId(),
+                playerControllers.stream().map(PlayerController::playerId).toList());
             if (room == null) {
                 log.debug("将玩家从房间中移除失败 gameType = {},roomId = {},playerIds = {}",
-                        this.room.getRoomCfgId(),
-                        this.room.getId(),
-                        playerControllers.stream().map(PlayerController::playerId).map(String::valueOf).collect(Collectors.joining(",")));
+                    this.room.getRoomCfgId(),
+                    this.room.getId(),
+                    playerControllers.stream().map(PlayerController::playerId).map(String::valueOf).collect(Collectors.joining(",")));
                 result.code = Code.FAIL;
                 return result;
             }
@@ -709,7 +713,7 @@ public abstract class AbstractRoomController<RC extends RoomCfg, R extends Room>
         return roomState;
     }
 
-    public RoomPlayer getRoomPlayer(long playerId){
+    public RoomPlayer getRoomPlayer(long playerId) {
         return room.getRoomPlayers().get(playerId);
     }
 }
